moving utilities.py and transferring re-usable modules out of prison_population.py
